+++ controllers/__init__.py
@@ -1,3 +1,4 @@
+from .DASFController import *
+from .SFYGController import *
 from .SolefootController import *
 from .WheelfootController import *
-from .DASFController import *